requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.26.2

# Data acquisition
yfinance==0.2.33
akshare==1.12.88

# Technical indicators
ta==0.11.0
talib-binary==0.4.28

# Machine learning
scikit-learn==1.3.2
torch==2.1.2
pytorch-lightning==2.1.3

# Time series models
statsmodels==0.14.1
prophet==1.1.5

# Mamba model (state space model)
mamba-ssm==1.2.0
causal-conv1d==1.1.0

# Evaluation and visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utils
tqdm==4.66.1
pyyaml==6.0.1
joblib==1.3.2

# Jupyter (for exploration)
jupyter==1.0.0
ipykernel==6.27.1

